requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
chardet==2.0.1
colorama==0.2.5
Django==1.10.6
django-allauth==0.34.0
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-formset-js==0.5.0
django-jquery-js==3.1.1
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
dnspython==1.11.1
ephem==3.7.6.0
html5lib==0.999
kombu==4.1.0
oauthlib==2.0.6
pycrypto==2.6.1
PyJWT==1.5.3
PyMySQL==0.7.11
python-openid==2.2.5
pytz==2017.3
redis==2.10.6
requests==2.2.1
requests-oauthlib==0.8.0
six==1.5.2
urllib3==1.7.1
uWSGI==2.0.15
vine==1.1.4
